continued to add to text extraction workflow
Added some new stat collection on what goods are traded most, as well as the ability to break down those stats by which area they are from. Also finished a version of specific and kinds of goods lists that is used to determine what goods are traded by which towns.