added txt file with a list of festival non-dates
This list comes from looking for everything between an ordinal word and a comma, period, or semicolon, then a deletion of every result with a number in it. This should be used to make a list of holidays, both movable and fixed, to be looked up by a function to assign dates to them.