further work on integrating holidays into broader dataset
There has been significant progress getting the dates of holidays out. However, there are still numerous cases of the holidays getting mixed up or unnecessarily duplicated. Some of these seem to be differences of opinion on when a saint's day is celebrated, but others, particularly St. Nicholas the Wonderworker with his two holidays, remain ambiguous or unnecessarily duplicated.