used notebooks to help create goods lookup
goods_lookup.json is to be used for looking up goods. Surprisingly. The keys are what will be used for checking in the text, and those without spaces will check for various inflectedforms. The display key should be used for final output. The eng key is a rough translation, not sure what it's going to be used for, but it's helpful for me. The kind key indicates whether it's a general sort of good or a specific one, ie iron goods vs. raw iron.