work on added text processing.
Re-generated spreadsheets in nlp, and added new spreadsheets in output to push to tableau. In notebook, found ways to extract info on the level of goods, i.e. this good was traded on these dates at a place that has a local ID of x. This has proved very useful for visualization.