changes from interrogating dataset for new patterns
Looked for entries that didn't currently have a fair associated with them, and exported to a csv for examination by longest descriptions. This has led to some additions to holiday regex. The goal now is to include weekly fairs, but not to include fairs that have no possibility of a defined temporal element.